vfio: support CXL type-2 device in VFIO stub
vfio-cxl is introduced to support CXL type-2 device passthrough. It
provides necessary UABIs for QEMU to manipulate the CXL type-2 device:

- A new VFIO device flags to indicate the device is a CXL device.
- A new sub-region type to indicate the region is a CXL region.
- A new device cap will provide the necessary information QEMU needs.

To support the CXL type-2 device passthrough, QEMU needs to leverage the
new UABIs to enumerate the CXL device, acquire the CXL device info, and
mapping the CXL region.

Introduce the new flags, sub-region type, and device cap aligned with
vfio-cxl. Acquire the HDM offset and bar from the device cap. Add a write
notification handler to track VM writes on the CXL MMIO BAR. Update the
mapping between GPA in the virtual HDM decoder and CXL region when the VM
commits an HDM decoder.

Signed-off-by: Zhi Wang <zhiw@nvidia.com>